refactor(frank): create entrypoints, redirect consumers, reduce allowlist
